refactor: move PasswordConfirmationAction to Forms namespace
Moves the PasswordConfirmationAction from the root Filament Actions namespace to the Forms subdirectory to better organize the codebase and reflect its usage within forms. This change also updates the namespace references in Livewire components that use this action.
